stacks/search_app/requirements.txt
fastapi
uvicorn
pandas
pyarrow
duckdb
psycopg2-binary
python-dotenv
sqlalchemy
beautifulsoup4
lxml
